Create IntegrationTestFixtureBase class
